Move inputs to new package
