+++ src/elements/icon/package.json
@@ -13,6 +13,7 @@
   },
   "dependencies": {
     "@emotion/core": "^10.0.10",
+    "@uswitch/trustyle.styles": "^0.1.3",
     "react": "^16.7.0"
   }
 }
