+++ src/elements/icon/src/styles.ts
@@ -1,6 +1,6 @@
 import { css } from '@emotion/core'
 
-import { pxToRem } from '../../../styles/lib'
+import { pxToRem } from '@uswitch/trustyle.styles'
 
 import { Direction } from '.'
