+++ src/elements/input/package.json
@@ -1,6 +1,6 @@
 {
-  "name": "@uswitch/trustyle.text-input",
-  "version": "0.3.3",
+  "name": "@uswitch/trustyle.input",
+  "version": "0.1.0",
   "license": "MIT",
   "main": "lib/index.js",
   "publishConfig": {
